+++ irs-policy-store/src/main/java/org/eclipse/tractusx/irs/policystore/services/PolicyStoreService.java
@@ -56,6 +56,7 @@ import org.eclipse.tractusx.irs.policystore.exceptions.PolicyStoreException;
 import org.eclipse.tractusx.irs.policystore.models.CreatePolicyRequest;
 import org.eclipse.tractusx.irs.policystore.models.UpdatePolicyRequest;
 import org.eclipse.tractusx.irs.policystore.persistence.PolicyPersistence;
+import org.eclipse.tractusx.irs.policystore.validators.PolicyValidator;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.web.server.ResponseStatusException;
@@ -78,9 +79,6 @@ public class PolicyStoreService implements AcceptedPoliciesProvider {
 
     private final Clock clock;
 
-    private static final String MISSING_REQUEST_FIELD_MESSAGE =
-            "Request does not contain all required fields. " + "Missing: %s";
-
     private static final String DEFAULT = "default";
 
     /**
@@ -100,10 +98,12 @@ public class PolicyStoreService implements AcceptedPoliciesProvider {
 
     public PolicyStoreService(final DefaultAcceptedPoliciesConfig defaultAcceptedPoliciesConfig,
             final PolicyPersistence persistence, final EdcTransformer edcTransformer, final Clock clock) {
+
+        this.clock = clock;
+
         this.allowedPoliciesFromConfig = createDefaultPolicyFromConfig(defaultAcceptedPoliciesConfig);
         this.persistence = persistence;
         this.edcTransformer = edcTransformer;
-        this.clock = clock;
     }
 
     /**
@@ -126,32 +126,14 @@ public class PolicyStoreService implements AcceptedPoliciesProvider {
         return registeredPolicy;
     }
 
-    public Policy doRegisterPolicy(final Policy policy, final String businessPartnersNumber) {
-        validatePolicy(policy);
+    /* package */ Policy doRegisterPolicy(final Policy policy, final String businessPartnersNumber) {
+        PolicyValidator.validate(policy);
         policy.setCreatedOn(OffsetDateTime.now(clock));
         log.info("Registering new policy with id {}, valid until {}", policy.getPolicyId(), policy.getValidUntil());
         try {
             return persistence.save(businessPartnersNumber, policy);
         } catch (final PolicyStoreException e) {
             throw new ResponseStatusException(HttpStatus.BAD_REQUEST, e.getMessage(), e);
-        }
-    }
-
-    /**
-     * Checks whether policy from register policy request has all required fields
-     *
-     * @param policy policy to register
-     */
-    private void validatePolicy(final Policy policy) {
-
-        if (policy.getPermissions() == null) {
-            throw new ResponseStatusException(HttpStatus.BAD_REQUEST,
-                    String.format(MISSING_REQUEST_FIELD_MESSAGE, "odrl:permission"));
-        }
-
-        if (policy.getPermissions().stream().anyMatch(p -> p.getConstraint() == null)) {
-            throw new ResponseStatusException(HttpStatus.BAD_REQUEST,
-                    String.format(MISSING_REQUEST_FIELD_MESSAGE, "odrl:constraint"));
         }
     }
 
@@ -218,7 +200,7 @@ public class PolicyStoreService implements AcceptedPoliciesProvider {
         }
     }
 
-    private void deletePolicyForEachBpn(final String policyId, final List<String> bpnList) {
+    public void deletePolicyForEachBpn(final String policyId, final List<String> bpnList) {
         try {
             for (final String bpn : bpnList) {
                 persistence.delete(bpn, policyId);
@@ -321,11 +303,17 @@ public class PolicyStoreService implements AcceptedPoliciesProvider {
                                              new Constraint(acceptedPolicy.getLeftOperand(),
                                                      new Operator(OperatorType.fromValue(acceptedPolicy.getOperator())),
                                                      acceptedPolicy.getRightOperand())));
-        final Policy policy = new Policy(ConfiguredDefaultPolicy.DEFAULT_POLICY_ID, OffsetDateTime.now(),
-                OffsetDateTime.now().plusYears(ConfiguredDefaultPolicy.DEFAULT_POLICY_LIFETIME_YEARS),
-                List.of(new Permission(PolicyType.USE, new Constraints(constraints, constraints))));
 
-        return List.of(policy);
+        final OffsetDateTime now = OffsetDateTime.now(clock);
+        return List.of(Policy.builder()
+                             .policyId(ConfiguredDefaultPolicy.DEFAULT_POLICY_ID)
+                             .createdOn(now)
+                             .validUntil(now.plusYears(ConfiguredDefaultPolicy.DEFAULT_POLICY_LIFETIME_YEARS))
+                             .permissions(List.of(Permission.builder()
+                                                            .action(PolicyType.USE)
+                                                            .constraint(new Constraints(constraints, constraints))
+                                                            .build()))
+                             .build());
     }
 
 }