+++ irs-policy-store/src/test/java/org/eclipse/tractusx/irs/policystore/controllers/PolicyStoreControllerTest.java
@@ -24,25 +24,23 @@
 package org.eclipse.tractusx.irs.policystore.controllers;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.eclipse.tractusx.irs.edc.client.policy.ConstraintConstants.ACTIVE_MEMBERSHIP;
+import static org.eclipse.tractusx.irs.edc.client.policy.ConstraintConstants.FRAMEWORK_AGREEMENT_TRACEABILITY_ACTIVE;
+import static org.eclipse.tractusx.irs.edc.client.policy.ConstraintConstants.PURPOSE_ID_3_1_TRACE;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
-import java.io.StringReader;
 import java.time.OffsetDateTime;
 import java.util.Collections;
 import java.util.List;
 import java.util.Map;
+import java.util.UUID;
 import java.util.stream.Collectors;
 
-import jakarta.json.Json;
 import jakarta.json.JsonObject;
-import jakarta.json.JsonReader;
 import jakarta.servlet.http.HttpServletRequest;
-import org.eclipse.tractusx.irs.edc.client.policy.Constraint;
 import org.eclipse.tractusx.irs.edc.client.policy.Constraints;
-import org.eclipse.tractusx.irs.edc.client.policy.Operator;
-import org.eclipse.tractusx.irs.edc.client.policy.OperatorType;
 import org.eclipse.tractusx.irs.edc.client.policy.Permission;
 import org.eclipse.tractusx.irs.edc.client.policy.Policy;
 import org.eclipse.tractusx.irs.edc.client.policy.PolicyType;
@@ -51,6 +49,7 @@ import org.eclipse.tractusx.irs.policystore.models.CreatePolicyRequest;
 import org.eclipse.tractusx.irs.policystore.models.PolicyResponse;
 import org.eclipse.tractusx.irs.policystore.models.UpdatePolicyRequest;
 import org.eclipse.tractusx.irs.policystore.services.PolicyStoreService;
+import org.eclipse.tractusx.irs.policystore.testutil.PolicyStoreTestUtil;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
@@ -66,7 +65,7 @@ public class PolicyStoreControllerTest {
                 "@context": {
                     "odrl": "http://www.w3.org/ns/odrl/2/"
                 },
-                "@id": "policy-id",
+                "@id": "e917f5f-8dac-49ac-8d10-5b4d254d2b48",
                 "policy": {
                     "odrl:permission": [
                         {
@@ -112,10 +111,7 @@ public class PolicyStoreControllerTest {
         void registerAllowedPolicy() {
             // arrange
             final OffsetDateTime now = OffsetDateTime.now();
-            final JsonObject jsonObject;
-            try (JsonReader jsonReader = Json.createReader(new StringReader(REGISTER_POLICY_EXAMPLE_PAYLOAD))) {
-                jsonObject = jsonReader.readObject();
-            }
+            final JsonObject jsonObject = PolicyStoreTestUtil.toJsonObject(REGISTER_POLICY_EXAMPLE_PAYLOAD);
 
             // act
             final CreatePolicyRequest request = new CreatePolicyRequest(now.plusMinutes(1), null,
@@ -135,8 +131,13 @@ public class PolicyStoreControllerTest {
         @Test
         void getPolicies() {
             // arrange
-            final List<Policy> policies = List.of(
-                    new Policy("testId", OffsetDateTime.now(), OffsetDateTime.now(), createPermissions()));
+            final String policyId = randomPolicyId();
+            final List<Policy> policies = List.of(Policy.builder()
+                                                        .policyId(policyId)
+                                                        .createdOn(OffsetDateTime.now())
+                                                        .validUntil(OffsetDateTime.now())
+                                                        .permissions(createPermissions())
+                                                        .build());
             final String bpn = "bpn1";
             when(policyStoreServiceMock.getPolicies(List.of(bpn))).thenReturn(Map.of(bpn, policies));
 
@@ -151,8 +152,13 @@ public class PolicyStoreControllerTest {
         @Test
         void getAllPolicies() {
             // arrange
-            final List<Policy> policies = List.of(
-                    new Policy("testId", OffsetDateTime.now(), OffsetDateTime.now(), createPermissions()));
+            final String policyId = randomPolicyId();
+            final List<Policy> policies = List.of(Policy.builder()
+                                                        .policyId(policyId)
+                                                        .createdOn(OffsetDateTime.now())
+                                                        .validUntil(OffsetDateTime.now())
+                                                        .permissions(createPermissions())
+                                                        .build());
             final String bpn = "bpn1";
             when(policyStoreServiceMock.getPolicies(null)).thenReturn(Map.of(bpn, policies));
 
@@ -209,8 +215,11 @@ public class PolicyStoreControllerTest {
 
     private Constraints createConstraints() {
         return new Constraints(Collections.emptyList(),
-                List.of(new Constraint("Membership", new Operator(OperatorType.EQ), "active"),
-                        new Constraint("FrameworkAgreement.traceability", new Operator(OperatorType.EQ), "active"),
-                        new Constraint("PURPOSE", new Operator(OperatorType.EQ), "ID 3.1 Trace")));
+                List.of(ACTIVE_MEMBERSHIP, FRAMEWORK_AGREEMENT_TRACEABILITY_ACTIVE, PURPOSE_ID_3_1_TRACE));
     }
+
+    private static String randomPolicyId() {
+        return UUID.randomUUID().toString();
+    }
+
 }