+++ CHANGELOG.md
@@ -7,17 +7,40 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 _**For better traceability add the corresponding GitHub issue number in each changelog entry, please.**_
 
 ## [Unreleased]
+ 
+### Changed
+
+- Improved policy store API input validation. #528
+
+
+
+## Added
+
+- Added endpoint for dedicated removal of policy from BPNL. #559 
+- Integration Test Policy Store API Unhappy Path. #519
+
+## [5.1.2] - 2024-05-13
+
+### Fixed
+
+-  Cleaning up BPNLs without policies. #533
+
+### Changed
+
+- Updated default accepted policy to latest traceability framework agreement #596
+- BPN summary was removed from Job response #568
+- Shells in Job response will contain all submodel descriptors returned by provider, instead filtered by aspect-type parameter. #510
+
 
 ## [5.1.1] - 2024-05-08
 
 ### Fixed
 
-- Fixed issue in EDR Token renewal #358
+- Fixed issue in EDR token renewal. #358
 
 ### Added
 
 - Cucumber test step definitions for Policy Store API (Happy Path) including some test helper utilities. #518
- 
 
 ## [5.1.0] - 2024-05-06
 
@@ -33,8 +56,6 @@ _**For better traceability add the corresponding GitHub issue number in each cha
 
 - Update bouncycastle to 1.78 to fix CVE's.
 - Fixed validation of json-schemas - IRS is creating tombstone instead collecting Submodel payload, when it not passes validation of schema #522
-
-
 
 ## [5.0.0] - 2024-04-16
 
@@ -644,7 +665,9 @@ _**For better traceability add the corresponding GitHub issue number in each cha
 ### Unresolved
 - **Select Aspects you need**  You are able to select the needed aspects for which you want to collect the correct endpoint information.
 
-[Unreleased]: https://github.com/eclipse-tractusx/item-relationship-service/compare/5.1.1...HEAD
+
+[Unreleased]: https://github.com/eclipse-tractusx/item-relationship-service/compare/5.1.2...HEAD
+[5.1.2]: https://github.com/eclipse-tractusx/item-relationship-service/compare/5.1.1...5.1.2
 [5.1.1]: https://github.com/eclipse-tractusx/item-relationship-service/compare/5.1.0...5.1.1
 [5.1.0]: https://github.com/eclipse-tractusx/item-relationship-service/compare/5.0.0...5.1.0
 [5.0.0]: https://github.com/eclipse-tractusx/item-relationship-service/compare/4.9.0...5.0.0